infra/aws/lambda/image_optimizer/requirements.txt
Pillow==11.1.0
boto3==1.39.11
